feat(component): rework page header for better reusability - split into discreet components - added class overrides
